src/requirements.txt
matplotlib==3.9.3
numpy==1.26.4
tensorflow-probability==0.18.0
tensorflow-model-optimization==0.7.3
## The following requirements were added by pip freeze:
absl-py==2.1.0
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
fonttools==4.55.1
gast==0.6.0
kiwisolver==1.4.7
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
